@@ -1,6 +1,6 @@
 {
   "name": "ozwillo-dc-exporter",
-  "version": "0.12.0",
+  "version": "0.13.0",
   "description": "Ozwillo datacore exporter",
   "private": true,
   "scripts": {
@@ -25,51 +25,54 @@
     "autoprefixer": "^6.3.7",
     "babel-core": "^6.26.3",
     "babel-eslint": "^6.1.2",
-    "babel-loader": "^6.0.1",
+    "babel-loader": "^7.1.5",
     "babel-preset-es2015": "^6.9.0",
-    "babel-preset-react": "^6.11.1",
-    "babel-preset-stage-0": "^6.22.0",
-    "bootstrap-loader": "^1.1.0",
+    "babel-preset-react": "^6.24.1",
+    "babel-preset-stage-0": "^6.24.0",
+    "bootstrap-loader": "^3.0.1",
     "clean-webpack-plugin": "^0.1.19",
-    "css-loader": "^0.23.1",
+    "css-loader": "^1.0.0",
     "eslint": "^3.2.0",
     "eslint-plugin-babel": "^3.3.0",
     "eslint-plugin-react": "^5.2.2",
-    "extract-text-webpack-plugin": "^1.0.1",
     "file-loader": "^0.9.0",
     "imports-loader": "^0.6.5",
     "less": "^2.7.1",
-    "node-sass": "^3.8.0",
-    "postcss-loader": "^0.9.1",
-    "react-hot-loader": "^1.3.0",
+    "mini-css-extract-plugin": "0.4.2",
+    "node-sass": "^4.9.3",
+    "postcss-loader": "^3.0.0",
+    "react-hot-loader": "^4.3.0",
     "resolve-url-loader": "^1.6.0",
-    "sass-loader": "^4.0.0",
+    "sass-loader": "^7.1.0",
     "script-loader": "^0.7.0",
     "style-loader": "^0.13.0",
     "url-loader": "^0.5.7",
-    "webpack": "^1.13.1",
-    "webpack-dev-server": "^1.14.1",
-    "webpack-merge": "^0.14.1"
+    "webpack": "^4.17.0",
+    "webpack-cli": "3.1.0",
+    "webpack-dev-server": "^3.1.5",
+    "webpack-merge": "^4.1.4"
   },
   "dependencies": {
     "bootstrap": "^4.1.2",
     "bootstrap-sass": "^3.3.7",
+    "create-react-class": "^15.6.3",
     "date-fns": "^1.29.0",
     "debounce": "^1.0.0",
-    "es6-promise": "^3.2.1",
+    "es6-promise": "^4.2.0",
     "exports-loader": "^0.6.3",
-    "font-awesome": "^4.6.3",
-    "i18next": "^7.1.3",
+    "font-awesome": "^4.7.0",
+    "i18next": "^11.6.0",
     "i18next-browser-languagedetector": "^1.0.1",
     "i18next-xhr-backend": "^1.5.1",
     "imports-loader": "^0.6.5",
     "jquery": "^3.3.1",
     "popper.js": "^1.14.3",
-    "react": "^15.3.0",
-    "react-autosuggest": "^8.0.0",
-    "react-dom": "^15.3.0",
-    "react-i18next": "^2.2.3",
-    "react-router": "^2.6.1",
+    "react": "^16.4.0",
+    "react-autosuggest": "^9.4.0",
+    "react-dom": "^16.4.0",
+    "react-i18next": "^7.10.1",
+    "react-router": "^4.3.1",
+    "react-router-dom": "^4.3.1",
     "render-if": "^0.1.0",
     "whatwg-fetch": "^1.0.0"
   }
