@@ -3,11 +3,11 @@ package org.ozwillo.dcexporter;
 import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.springframework.test.context.web.WebAppConfiguration;
-import org.springframework.boot.test.SpringApplicationConfiguration;
+import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.test.context.junit4.SpringJUnit4ClassRunner;
 
 @RunWith(SpringJUnit4ClassRunner.class)
-@SpringApplicationConfiguration(classes = OzwilloDcExporterApplication.class)
+@SpringBootTest(classes = OzwilloDcExporterApplication.class)
 @WebAppConfiguration
 public class OzwilloDcExporterApplicationTests {
 
